+++ ocaml-lsp-server/src/testing.ml
@@ -1,6 +1,7 @@
 (**WARNING: This is for internal use in testing only *)
 
 module Compl = Compl
+module Destruct_line = Destruct_line_search
 module Document_symbol = Document_symbol
 module Merlin_kernel = Merlin_kernel
 module Prefix_parser = Prefix_parser